Moved metricsexporter servicers to servicers/protected
